refactor(service/cache/redis): turn userID to int64

+++ module/cache.go
@@ -3,8 +3,8 @@ package module
 // Cache : distributed shared cache module
 type Cache interface {
 	Ping() error
-	UpdateToken(userID int) (string, error)
-	GetUserIDByToken(token string) (int, error)
-	DelTokenByUserID(userID int) error
+	UpdateToken(userID int64) (string, error)
+	GetUserIDByToken(token string) (int64, error)
+	DelTokenByUserID(userID int64) error
 	DelTokenByToken(token string) error
 }
